+++ OTLib/OTEnvelope.h
@@ -325,9 +325,9 @@ class OTSymmetricKey
 private:
     bool            m_bIsGenerated;     // GetKey asserts if this is false; GenerateKey asserts if it's true.
     // ---------------------------------------------
-    unsigned int    m_nKeySize;         // The size, in bits. For example, 128 bit key, 256 bit key, etc.
+    uint16_t        m_nKeySize;         // The size, in bits. For example, 128 bit key, 256 bit key, etc.
     // ---------------------------------------------
-    int             m_nIterationCount;  // Stores the iteration count, which should probably be at least 2000. (Number of iterations used while generating key from passphrase.)
+    uint32_t        m_nIterationCount;  // Stores the iteration count, which should probably be at least 2000. (Number of iterations used while generating key from passphrase.)
     // ---------------------------------------------
 	OTPayload       m_dataSalt;         // Stores the SALT (which is used with the password for generating / retrieving the key from m_dataEncryptedKey)
 	OTPayload       m_dataIV;           // Stores the IV used internally for encrypting / decrypting the actual key (using the derived key) from m_dataEncryptedKey.
